==> src/6_all_experiment_results.ipynb <==
# Plot all experiment results

This notebook contains the comparison of all the results obtained with different closure strategies from the experiments in the thesis.

import pandas as pd
import matplotlib.pyplot as plt
from matplotlib.legend_handler import HandlerTuple
import numpy as np
import sumolib

import seaborn as sns
pal = sns.color_palette('colorblind')
print(pal.as_hex())
pal.as_hex()

col = {
    'co2': '#0173b2',
    'hf': '#de8f05',
    'he': '#029e73',
    'lf': '#d55e00',
    'le': '#949494',
    'mix': '#ca9161'
}

#### parameters

city = 'Milano_big'
exp_values = ['100', '90', '80', '70', '60', '50', '40', '30', '20', '10', '1']
exp_names = ['top'+v for v in exp_values]
exp_rands = ['rand'+v for v in exp_values]
exp_names_HF = ['HF_top'+v for v in exp_values]
exp_names_HE = ['HE_top'+v for v in exp_values]
exp_names_LF = ['LF_top'+v for v in exp_values]
exp_names_LE = ['LE_top'+v for v in exp_values]
exp_names_MIX = ['MIX_top'+v for v in exp_values[:-1]+['4']]
#exp_intervals = ['top'+v for v in ['20-25', '25-30', '30-35', '35-40']]

# road net
road_network_path = "../data/road_net/"+city+"/"+city+"_road_network.net.xml"

# folder result paths
results_path = ['../data/simulations/'+city+'/'+n+'/results/'+n+'_total_co2.csv' for n in exp_names]
stats_path = ['../data/simulations/'+city+'/'+n+'/results/'+n+'_stats.csv' for n in exp_names]
results_path_HF = ['../data/simulations/'+city+'/'+n+'/results/'+n+'_total_co2.csv' for n in exp_names_HF]
stats_path_HF = ['../data/simulations/'+city+'/'+n+'/results/'+n+'_stats.csv' for n in exp_names_HF]
results_path_HE = ['../data/simulations/'+city+'/'+n+'/results/'+n+'_total_co2.csv' for n in exp_names_HE]
stats_path_HE = ['../data/simulations/'+city+'/'+n+'/results/'+n+'_stats.csv' for n in exp_names_HE]
results_path_LF = ['../data/simulations/'+city+'/'+n+'/results/'+n+'_total_co2.csv' for n in exp_names_LF]
stats_path_LF = ['../data/simulations/'+city+'/'+n+'/results/'+n+'_stats.csv' for n in exp_names_LF]
results_path_LE = ['../data/simulations/'+city+'/'+n+'/results/'+n+'_total_co2.csv' for n in exp_names_LE]
stats_path_LE = ['../data/simulations/'+city+'/'+n+'/results/'+n+'_stats.csv' for n in exp_names_LE]
results_path_MIX = ['../data/simulations/'+city+'/'+n+'/results/'+n+'_total_co2.csv' for n in exp_names_MIX]
stats_path_MIX = ['../data/simulations/'+city+'/'+n+'/results/'+n+'_stats.csv' for n in exp_names_MIX]
#results_intervals_path = ['../data/simulations/'+city+'/'+n+'/results/'+n+'_total_co2.csv' for n in exp_intervals]

# folder for saving plots
plots_path = '../data/plots/'

### Aggregate all csv

road_network = sumolib.net.readNet(road_network_path, withInternal=False)
n_edges = len(road_network.getEdges())

def aggregate_csv(paths):
    l = []

    for res_csv in paths:
        df_tmp = pd.read_csv(res_csv)
        l.append(df_tmp)

    df = pd.concat(l, ignore_index=True)

    return df

res = {
    'co2': aggregate_csv(results_path),
    'hf': aggregate_csv(results_path_HF),
    'he': aggregate_csv(results_path_HE),
    'lf': aggregate_csv(results_path_LF),
    'le': aggregate_csv(results_path_LE),
    'mix': aggregate_csv(results_path_MIX)
}

stat = {
    'co2': aggregate_csv(stats_path),
    'hf': aggregate_csv(stats_path_HF),
    'he': aggregate_csv(stats_path_HE),
    'lf': aggregate_csv(stats_path_LF),
    'le': aggregate_csv(stats_path_LE),
    'mix': aggregate_csv(stats_path_MIX)
}

for i, df in res.items():
    df.drop_duplicates(inplace=True)

def get_exp_df(df, colname, exp_name):
    return df[df[colname].str.startswith(exp_name)].copy()

res['baseline'] = get_exp_df(res['co2'], 'experiment', 'baseline')

for i, df in res.items():
    if i == 'co2':
        res[i] = {
            'top': get_exp_df(df, 'experiment', 'top'),
            'rand': get_exp_df(df, 'experiment', 'rand')
        }
    elif i == 'hf':
        res[i] = {
            'top': get_exp_df(df, 'experiment', 'HF_top'),
            'rand': get_exp_df(df, 'experiment', 'rand')
        }
    elif i == 'he':
        res[i] = {
            'top': get_exp_df(df, 'experiment', 'HE_top'),
            'rand': get_exp_df(df, 'experiment', 'rand')
        }
    elif i == 'lf':
        res[i] = {
            'top': get_exp_df(df, 'experiment', 'LF_top'),
            'rand': get_exp_df(df, 'experiment', 'rand')
        }
    elif i == 'le':
        res[i] = {
            'top': get_exp_df(df, 'experiment', 'LE_top'),
            'rand': get_exp_df(df, 'experiment', 'rand')
        }
    elif i == 'mix':
        res[i] = {
            'top': get_exp_df(df, 'experiment', 'MIX_top'),
            'rand': get_exp_df(df, 'experiment', 'rand')
        }

for i, df in stat.items():
    if i == 'co2':
        stat[i] = {
            'top': get_exp_df(df, 'exp', 'top'),
            'rand': get_exp_df(df, 'exp', 'rand')
        }
    elif i == 'hf':
        stat[i] = {
            'top': get_exp_df(df, 'exp', 'HF_top'),
            'rand': get_exp_df(df, 'exp', 'rand')
        }
    elif i == 'he':
        stat[i] = {
            'top': get_exp_df(df, 'exp', 'HE_top'),
            'rand': get_exp_df(df, 'exp', 'rand')
        }
    elif i == 'lf':
        stat[i] = {
            'top': get_exp_df(df, 'exp', 'LF_top'),
            'rand': get_exp_df(df, 'exp', 'rand')
        }
    elif i == 'le':
        stat[i] = {
            'top': get_exp_df(df, 'exp', 'LE_top'),
            'rand': get_exp_df(df, 'exp', 'rand')
        }
    elif i == 'mix':
        stat[i] = {
            'top': get_exp_df(df, 'exp', 'MIX_top'),
            'rand': get_exp_df(df, 'exp', 'rand')
        }

def random_aggregator(df):
    df = df.groupby(by='removed_road').agg({'removed_edge': 'mean'}).reset_index()
    df['removed_edge'] = df['removed_edge'].round()
    return df

for i, d in stat.items():
    stat[i]['rand'] = random_aggregator(d['rand'])

def replace_sort(df, str2replace):
    df['removed_roads'] = df['experiment'].replace({str2replace: '', '_norm':''}, regex=True).astype(int)
    df.sort_values(by=['removed_roads'], ascending=True, inplace=True)
    return df

res['co2']['top']

for i, d in res.items():
    if i == 'co2':
        res[i]['top'] = replace_sort(d['top'], 'top')
        res[i]['rand'] = replace_sort(d['rand'], 'rand')
    elif i == 'hf':
        res[i]['top'] = replace_sort(d['top'], 'HF_top')
        res[i]['rand'] = replace_sort(d['rand'], 'rand')
    elif i == 'he':
        res[i]['top'] = replace_sort(d['top'], 'HE_top')
        res[i]['rand'] = replace_sort(d['rand'], 'rand')
    elif i == 'lf':
        res[i]['top'] = replace_sort(d['top'], 'LF_top')
        res[i]['rand'] = replace_sort(d['rand'], 'rand')
    elif i == 'le':
        res[i]['top'] = replace_sort(d['top'], 'LE_top')
        res[i]['rand'] = replace_sort(d['rand'], 'rand')
    elif i == 'mix':
        res[i]['top'] = replace_sort(d['top'], 'MIX_top')
        res[i]['rand'] = replace_sort(d['rand'], 'rand')

res['co2']['top'].head()

def merge_stats(df, df_stats):
    df_stats = df_stats.merge(df, left_on='removed_road', right_on='removed_roads')[['removed_roads', 'removed_edge', 'edges_mean']]
    df_stats['fraction'] = df_stats['edges_mean']/(n_edges-df_stats['removed_edge'])
    return df_stats

for i, d in stat.items():
    stat[i]['top'] = merge_stats(res[i]['top'], d['top'])
    stat[i]['rand'] = merge_stats(res[i]['rand'], d['rand'])

### Plot

CO2 emissions for all experiments in unique plot

plt.figure(figsize=(6, 4))

# data
## CO2 exp
tmp = res['co2']['top']
plt.errorbar(tmp['removed_roads'], tmp['co2_mean'], yerr=tmp['co2_std'], label='Top CO2',
            c=col['co2'], marker='*', ms=8, capsize=0, elinewidth=1)
tmp = res['co2']['rand']
plt.errorbar(tmp['removed_roads'], tmp['co2_mean'], yerr=tmp['co2_std'], label='Rand CO2',
            c=col['co2'], linestyle='--', marker='*', ms=8, capsize=0, elinewidth=1)

## HF
tmp = res['hf']['top']
plt.errorbar(tmp['removed_roads'], tmp['co2_mean'], yerr=tmp['co2_std'], label='Top high k-road, high VOC',
            c=col['hf'], marker='.', ms=8, capsize=0, elinewidth=1)
tmp = res['hf']['rand']
plt.errorbar(tmp['removed_roads'], tmp['co2_mean'], yerr=tmp['co2_std'], label='Rand high k-road, high VOC',
            c=col['hf'], linestyle='--', marker='.', ms=8, capsize=0, elinewidth=1)

## HE
tmp = res['he']['top']
plt.errorbar(tmp['removed_roads'], tmp['co2_mean'], yerr=tmp['co2_std'], label='Top high k-road, low VOC',
            c=col['he'], marker='^', ms=6, capsize=0, elinewidth=1)
tmp = res['he']['rand']
plt.errorbar(tmp['removed_roads'], tmp['co2_mean'], yerr=tmp['co2_std'], label='Rand high k-road, low VOC',
            c=col['he'], linestyle='--', marker='^', ms=6, capsize=0, elinewidth=1)

## LF
tmp = res['lf']['top']
plt.errorbar(tmp['removed_roads'], tmp['co2_mean'], yerr=tmp['co2_std'], label='Top low k-road, high VOC',
            c=col['lf'], marker='s', ms=6, capsize=0, elinewidth=1)
tmp = res['lf']['rand']
plt.errorbar(tmp['removed_roads'], tmp['co2_mean'], yerr=tmp['co2_std'], label='Rand low k-road, high VOC',
            c=col['lf'], linestyle='--', marker='s', ms=6, capsize=0, elinewidth=1)

## LE
tmp = res['le']['top']
plt.errorbar(tmp['removed_roads'], tmp['co2_mean'], yerr=tmp['co2_std'], label='Top low k-road, low VOC',
            c=col['le'], marker='d', ms=6, capsize=0, elinewidth=1)
tmp = res['le']['rand']
plt.errorbar(tmp['removed_roads'], tmp['co2_mean'], yerr=tmp['co2_std'], label='Rand low k-road, low VOC',
            c=col['le'], linestyle='--', marker='d', ms=6, capsize=0, elinewidth=1)

## MIX
tmp = res['mix']['top']
plt.errorbar(tmp['removed_roads'], tmp['co2_mean'], yerr=tmp['co2_std'], label='Mixed strategies',
            c=col['mix'], marker='h', ms=6, capsize=0, elinewidth=1)
tmp = res['mix']['rand']
plt.errorbar(tmp['removed_roads'], tmp['co2_mean'], yerr=tmp['co2_std'], label='Random strategies',
            c=col['mix'], linestyle='--', marker='h', ms=6, capsize=0, elinewidth=1)

## baseline
tmp = res['baseline']
b = plt.axhline(tmp['co2_mean'].values, color='black', linestyle='--', label='baseline')
plt.axhspan(tmp['co2_mean'].values+tmp['co2_std'].values, tmp['co2_mean'].values-tmp['co2_std'].values, facecolor='black', alpha=0.1)


# decorators
plt.title('Total CO2 emission for each removal experiment')
plt.ylabel('Total CO2 (mg)')
plt.xlabel('#removed roads')
plt.xticks([int(v) for v in exp_values])
plt.yticks(np.arange(8.5e+10, 1.5e+11, step=5e+9))
plt.grid()
plt.legend(loc='upper left', bbox_to_anchor=(1.01,1.0))

plt.savefig(plots_path+'Milano_big.png', bbox_inches ="tight")
plt.show()

CO2 emissions obtained from infomred strategies in unique plot

plt.figure(figsize=(6, 4))

# data
## CO2 exp
tmp = res['co2']['top']
plt.errorbar(tmp['removed_roads'], tmp['co2_mean'], yerr=tmp['co2_std'], label='Top CO2',
            c=col['co2'], marker='*', ms=8, capsize=0, elinewidth=1)

## HF
tmp = res['hf']['top']
plt.errorbar(tmp['removed_roads'], tmp['co2_mean'], yerr=tmp['co2_std'], label='Top high k-road, high VOC',
            c=col['hf'], marker='.', ms=8, capsize=0, elinewidth=1)

## HE
tmp = res['he']['top']
plt.errorbar(tmp['removed_roads'], tmp['co2_mean'], yerr=tmp['co2_std'], label='Top high k-road, low VOC',
            c=col['he'], marker='^', ms=6, capsize=0, elinewidth=1)

## LF
tmp = res['lf']['top']
plt.errorbar(tmp['removed_roads'], tmp['co2_mean'], yerr=tmp['co2_std'], label='Top low k-road, high VOC',
            c=col['lf'], marker='s', ms=6, capsize=0, elinewidth=1)

## LF
tmp = res['le']['top']
plt.errorbar(tmp['removed_roads'], tmp['co2_mean'], yerr=tmp['co2_std'], label='Top low k-road, low VOC',
            c=col['le'], marker='d', ms=6, capsize=0, elinewidth=1)

## MIX
tmp = res['mix']['top']
plt.errorbar(tmp['removed_roads'], tmp['co2_mean'], yerr=tmp['co2_std'], label='Mixed strategies',
            c=col['mix'], marker='h', ms=6, capsize=0, elinewidth=1)

# baseline
tmp = res['baseline']
b = plt.axhline(tmp['co2_mean'].values, color='black', linestyle='--', label='baseline')
plt.axhspan(tmp['co2_mean'].values+tmp['co2_std'].values, tmp['co2_mean'].values-tmp['co2_std'].values, facecolor='black', alpha=0.1)


# decorators
plt.title('Total CO2 emission for each removal experiment')
plt.ylabel('Total CO2 (mg)')
plt.xlabel('#removed roads')
plt.xticks([int(v) for v in exp_values])
plt.yticks(np.arange(8.5e+10, 1.5e+11, step=5e+9))
plt.grid()
plt.legend(loc='upper left', bbox_to_anchor=(1.01,1.0))

plt.savefig(plots_path+'Milano_big_top_exp.png', bbox_inches ="tight")
plt.show()

Comparison of CO2 emissions obtained from informed and uninformed strategies, in separate plots

fig, axs = plt.subplots(3, 2, figsize=(12,15))

# data
## CO2 exp
tmp = res['co2']['top']
axs[0,0].errorbar(tmp['removed_roads'], tmp['co2_mean'], yerr=tmp['co2_std'], label='Top',
            c=col['co2'], marker='*', ms=8, capsize=0, elinewidth=1)
tmp = res['co2']['rand']
axs[0,0].errorbar(tmp['removed_roads'], tmp['co2_mean'], yerr=tmp['co2_std'], label='Rand',
            c=col['co2'], linestyle='--', marker='*', ms=8, capsize=0, elinewidth=1)
axs[0,0].set_title('Removal by CO2')

## HF
tmp = res['hf']['top']
axs[0,1].errorbar(tmp['removed_roads'], tmp['co2_mean'], yerr=tmp['co2_std'], label='Top',
            c=col['hf'], marker='.', ms=8, capsize=0, elinewidth=1)
tmp = res['hf']['rand']
axs[0,1].errorbar(tmp['removed_roads'], tmp['co2_mean'], yerr=tmp['co2_std'], label='Rand',
            c=col['hf'], linestyle='--', marker='.', ms=8, capsize=0, elinewidth=1)
axs[0,1].set_title('Removal by high k-road, high VOC')

## HE
tmp = res['he']['top']
axs[1,0].errorbar(tmp['removed_roads'], tmp['co2_mean'], yerr=tmp['co2_std'], label='Top',
            c=col['he'], marker='^', ms=6, capsize=0, elinewidth=1)
tmp = res['he']['rand']
axs[1,0].errorbar(tmp['removed_roads'], tmp['co2_mean'], yerr=tmp['co2_std'], label='Rand',
            c=col['he'], linestyle='--', marker='^', ms=6, capsize=0, elinewidth=1)
axs[1,0].set_title('Removal by high k-road, low VOC')

## LF
tmp = res['lf']['top']
axs[1,1].errorbar(tmp['removed_roads'], tmp['co2_mean'], yerr=tmp['co2_std'], label='Top',
            c=col['lf'], marker='s', ms=6, capsize=0, elinewidth=1)
tmp = res['lf']['rand']
axs[1,1].errorbar(tmp['removed_roads'], tmp['co2_mean'], yerr=tmp['co2_std'], label='Rand',
            c=col['lf'], linestyle='--', marker='s', ms=6, capsize=0, elinewidth=1)
axs[1,1].set_title('Removal by low k-road, high VOC')

## LE
tmp = res['le']['top']
axs[2,0].errorbar(tmp['removed_roads'], tmp['co2_mean'], yerr=tmp['co2_std'], label='Top',
            c=col['le'], marker='d', ms=6, capsize=0, elinewidth=1)
tmp = res['le']['rand']
axs[2,0].errorbar(tmp['removed_roads'], tmp['co2_mean'], yerr=tmp['co2_std'], label='Rand',
            c=col['le'], linestyle='--', marker='d', ms=6, capsize=0, elinewidth=1)
axs[2,0].set_title('Removal by low k-road, low VOC')

## LE
tmp = res['mix']['top']
axs[2,1].errorbar(tmp['removed_roads'], tmp['co2_mean'], yerr=tmp['co2_std'], label='Top',
            c=col['mix'], marker='h', ms=6, capsize=0, elinewidth=1)
tmp = res['mix']['rand']
axs[2,1].errorbar(tmp['removed_roads'], tmp['co2_mean'], yerr=tmp['co2_std'], label='Rand',
            c=col['mix'], linestyle='--', marker='h', ms=6, capsize=0, elinewidth=1)
axs[2,1].set_title('Removal with mixed strategy')


# baseline
tmp = res['baseline']
for ax in axs.reshape(-1):
    b = ax.axhline(tmp['co2_mean'].values, color='black', linestyle='--', label='baseline')
    ax.axhspan(tmp['co2_mean'].values+tmp['co2_std'].values, tmp['co2_mean'].values-tmp['co2_std'].values, facecolor='black', alpha=0.1)
    ax.set_ylabel('Total CO2 (mg)')
    ax.set_xlabel('#removed roads')
    ax.set_xticks([int(v) for v in exp_values])
    ax.set_yticks(np.arange(8.5e+10, 1.5e+11, step=5e+9))
    ax.grid()
    ax.legend(loc='upper left')

#axs[2,1].remove()

plt.savefig(plots_path+'Milano_big_clustered.png', bbox_inches ="tight")
plt.show()

Percentage of traveled edges for all experiments in unique plot

plt.figure(figsize=(6, 4))

## CO2
tmp = stat['co2']['top']
plt.errorbar(tmp['removed_roads'], tmp['fraction']*100, label='top CO2',
             c=col['co2'], marker='*', ms=8)
tmp = stat['co2']['rand']
plt.errorbar(tmp['removed_roads'], tmp['fraction']*100, label='rand CO2',
             c=col['co2'], linestyle='--', marker='*', ms=8)

## HF
tmp = stat['hf']['top']
plt.errorbar(tmp['removed_roads'], tmp['fraction']*100, label='top_HF',
             c=col['hf'], marker='.', ms=8)
tmp = stat['hf']['rand']
plt.errorbar(tmp['removed_roads'], tmp['fraction']*100, label='rand_HF',
             c=col['hf'], linestyle='--', marker='.', ms=8)

## HE
tmp = stat['he']['top']
plt.errorbar(tmp['removed_roads'], tmp['fraction']*100, label='top_HE',
             c=col['he'], marker='^', ms=6)
tmp = stat['he']['rand']
plt.errorbar(tmp['removed_roads'], tmp['fraction']*100, label='rand_HE',
             c=col['he'], linestyle='--', marker='^', ms=6)

## LF
tmp = stat['lf']['top']
plt.errorbar(tmp['removed_roads'], tmp['fraction']*100, label='top_LF',
             c=col['lf'], marker='s', ms=6)
tmp = stat['lf']['rand']
plt.errorbar(tmp['removed_roads'], tmp['fraction']*100, label='rand_LF',
             c=col['lf'], linestyle='--', marker='s', ms=6)

## LE
tmp = stat['le']['top']
plt.errorbar(tmp['removed_roads'], tmp['fraction']*100, label='top_LE',
             c=col['le'], marker='d', ms=6)
tmp = stat['le']['rand']
plt.errorbar(tmp['removed_roads'], tmp['fraction']*100, label='rand_LE',
             c=col['le'], linestyle='--', marker='d', ms=6)

## MIX
tmp = stat['mix']['top']
plt.errorbar(tmp['removed_roads'], tmp['fraction']*100, label='top_MIX',
             c=col['mix'], marker='h', ms=6)
tmp = stat['mix']['rand']
plt.errorbar(tmp['removed_roads'], tmp['fraction']*100, label='rand_MIX',
             c=col['mix'], linestyle='--', marker='h', ms=6)

# baseline
tmp = res['baseline']
b = plt.axhline(tmp['edges_mean'].values/n_edges*100, color='black', linestyle='--', label='baseline')


# decorators
plt.title('Percentage of travelled edges over total available edges')
plt.ylabel('% of travelled edges')
plt.xlabel('#removed roads')
plt.xticks([int(v) for v in exp_values])
#plt.yticks(np.arange(8.4e+10, 1.3e+11, step=4e+9))
plt.grid()
plt.legend(loc='upper left', bbox_to_anchor=(1.01,1.0))

plt.savefig(plots_path+'Milano_big_edges.png', bbox_inches ="tight")
plt.show()

Percentage of impacted vehicles for all experiments in unique plot

tot_v = 30000

plt.figure(figsize=(6, 4))

## CO2
tmp = res['co2']['top']
plt.errorbar(tmp['removed_roads'], tmp['impacted_v_mean']/tot_v*100, label='top CO2', 
            c=col['co2'], marker='*', ms=8)
tmp = res['co2']['rand']
plt.errorbar(tmp['removed_roads'], tmp['impacted_v_mean']/tot_v*100, label='rand CO2',
            c=col['co2'], linestyle='--', marker='*', ms=8)

## HF
tmp = res['hf']['top']
plt.errorbar(tmp['removed_roads'], tmp['impacted_v_mean']/tot_v*100, label='top_HF',
            c=col['hf'], marker='.', ms=8)
tmp = res['hf']['rand']
plt.errorbar(tmp['removed_roads'], tmp['impacted_v_mean']/tot_v*100, label='rand_HF',
            c=col['hf'], linestyle='--', marker='.', ms=8)

## HE
tmp = res['he']['top']
plt.errorbar(tmp['removed_roads'], tmp['impacted_v_mean']/tot_v*100, label='top_HE',
            c=col['he'], marker='^', ms=6)
tmp = res['he']['rand']
plt.errorbar(tmp['removed_roads'], tmp['impacted_v_mean']/tot_v*100, label='rand_HE',
            c=col['he'], linestyle='--', marker='^', ms=6)

## LF
tmp = res['lf']['top']
plt.errorbar(tmp['removed_roads'], tmp['impacted_v_mean']/tot_v*100, label='top_LF',
            c=col['lf'], marker='s', ms=6)
tmp = res['lf']['rand']
plt.errorbar(tmp['removed_roads'], tmp['impacted_v_mean']/tot_v*100, label='rand_LF',
            c=col['lf'], linestyle='--', marker='s', ms=6)

## LE
tmp = res['le']['top']
plt.errorbar(tmp['removed_roads'], tmp['impacted_v_mean']/tot_v*100, label='top_LE',
            c=col['le'], marker='d', ms=6)
tmp = res['le']['rand']
plt.errorbar(tmp['removed_roads'], tmp['impacted_v_mean']/tot_v*100, label='rand_LE',
            c=col['le'], linestyle='--', marker='d', ms=6)

## MIX
tmp = res['mix']['top']
plt.errorbar(tmp['removed_roads'], tmp['impacted_v_mean']/tot_v*100, label='top_MIX',
            c=col['mix'], marker='h', ms=6)
tmp = res['mix']['rand']
plt.errorbar(tmp['removed_roads'], tmp['impacted_v_mean']/tot_v*100, label='rand_MIX',
            c=col['mix'], linestyle='--', marker='h', ms=6)


# decorators
plt.title('Percentage of vehicles impacted by removal experiment')
plt.ylabel('% of impacted vehicles')
plt.xlabel('#removed roads')
plt.xticks([int(v) for v in exp_values])
plt.grid()
plt.legend(loc='upper left', bbox_to_anchor=(1.01,1.0))

plt.savefig(plots_path+'Milano_big_vehicles.png', bbox_inches ="tight")
plt.show()
